Changed some logics
