--- FactoryMethodTask/ShapeType.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace FactoryMethodtask
-{
-    public enum ShapeType
-    {
-        Circle, Triangle, Rectangle
-    }
-}
