+++ FactoryMethodTask/shapes/factories/bases/ShapeFactory.cs
@@ -1,0 +1,11 @@
+using System;
+using FactoryMethodTask.shapes.bases;
+
+namespace FactoryMethodTask.shapes.factories.bases
+{
+    public interface ShapeFactory
+    {
+        public IShape Shape { get; }
+    }
+}
+
